backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings>=2.0.0
google-genai>=0.3.0
supabase>=2.0.0
